tetra: uses grpc.NewClient in NewClientWithDefaultContext
Switch from grpc.DialContext to grpc.NewClient which does not perform
any I/O and wait lazily for the first RPC call to be made to connect.
This is the way to go as DialContext is getting deprecated and the gRPC
team has been writing that using the other is an anti-pattern.

https://github.com/grpc/grpc-go/blob/master/Documentation/anti-patterns.md

Signed-off-by: Mahe Tardy <mahe.tardy@gmail.com>