mbset: introduce the tg_mbset_map
The goal is to implement a 'followChildren: true' attribute in the
MatchBinaries operator that can be used to match not only processes with
the specified binary name, but also their children.

To do so, we need to mark processes at exec time. This patch introduces
tha bpf map that will be used for the marking.

The map has paths as keys and a bitset of ids as a value.

The map is written by user-space and read in BPF.
Each id corresponds to a MatchBinaries section with
'followChildren: true'. The mbset package implements the user-space side
of the map.

For example, consider the following sections:

- matchBinaries:
  - operator: "In"
    values:
    - "/usr/bin/bash"
    - "/usr/bin/dash"
    followChildren: true
ID: 0

- matchBinaries:
  - operator: "In"
    values:
    - "/usr/bin/bash"
    - "/usr/bin/csh"
    followChildren: true
ID: 5

The map then will look like:

/usr/bin/bash -> (1<<0) | (1<<5) = 0x21
/usr/bin/dash -> (1<<0)          = 0x1
/usr/bin/csh  -> (1<<5)          = 0x20

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>