testutils: improve loadGenericSensorTest
loadGenericSensorTest is a helper function to load a generic sensor.
With the addition of the tg_mbset_map in the previous commits, when
using a matchBinaries with followChildren: true because for the latter,
we need to base sensor to be loaded.

This fixes the above issue by initializing the base (exec) sensor
before running SensorsFromPolicy.

Signed-off-by: Kornilios Kourtis <kornilios@isovalent.com>